refactor(server): split tool_grammar::wire along its three shapes
`wire.rs` was 607 lines holding the JSON, element and harmony root
rules in one file, and the next change adds two more shapes to it. The
file-size rule says split first, so this is the split with no behaviour
change: `shape` and the shared helpers stay in `wire/mod.rs`, and each
root rule moves to a module named after its shape.

What each shape does about a value's TYPE moves with the shape that
decides it, rather than staying in one header that has to describe all
of them.
